# docs/api/openapi.m13.json
{
...
  "components": {
    "schemas": {
      "AgentSpecDto": {
        "additionalProperties": false,
        "description": "AgentSpec 实例（Role 的配置实例；可独立绑定模型）。",
        "properties": {
          "budget_policy_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Budget Policy Ref"
          },
          "capability_refs": {
            "items": {
              "type": "string"
            },
            "title": "Capability Refs",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "max_context_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Context Tokens"
          },
          "max_iterations": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Iterations"
          },
          "model_binding": {
            "additionalProperties": false,
            "title": "Model Binding",
            "type": "object"
          },
          "role": {
            "title": "Role",
            "type": "string"
          },
          "runtime_kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Runtime Kind"
          },
          "skill_refs": {
            "items": {
              "type": "string"
            },
            "title": "Skill Refs",
            "type": "array"
          },
          "workspace_policy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workspace Policy"
          }
        },
        "required": [
          "id",
          "role"
        ],
        "title": "AgentSpecDto",
        "type": "object"
      },
      "AgentUpdateDto": {
        "additionalProperties": false,
        "description": "per-Agent ModelBinding 更新（semantic 变更走 Manifest Revision/Fork）。",
        "properties": {
          "max_context_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Context Tokens"
          },
          "max_iterations": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Iterations"
          },
          "model_binding": {
            "anyOf": [
              {
                "additionalProperties": false,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Binding"
          },
          "workspace_policy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workspace Policy"
          }
        },
        "title": "AgentUpdateDto",
        "type": "object"
      },
...
      "CompileResultDto": {
        "additionalProperties": false,
        "properties": {
          "findings": {
            "items": {
              "additionalProperties": false,
              "type": "object"
            },
            "title": "Findings",
            "type": "array"
          },
          "plan_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan Id"
          },
          "protocol_digest": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Protocol Digest"
          },
          "successful": {
            "title": "Successful",
            "type": "boolean"
          }
        },
        "required": [
          "successful"
        ],
        "title": "CompileResultDto",
        "type": "object"
      },
...
      "DryRunProjectionDto": {
        "additionalProperties": false,
        "properties": {
          "agent_models": {
            "additionalProperties": false,
            "title": "Agent Models",
            "type": "object"
          },
          "approval_actions": {
            "items": {
              "type": "string"
            },
            "title": "Approval Actions",
            "type": "array"
          },
          "budget_reservations": {
            "items": {
              "additionalProperties": false,
              "type": "object"
            },
            "title": "Budget Reservations",
            "type": "array"
          },
          "compute_profiles": {
            "additionalProperties": false,
            "title": "Compute Profiles",
            "type": "object"
          },
          "estimated_cost_minor": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Cost Minor"
          },
          "role_counts": {
            "additionalProperties": false,
            "title": "Role Counts",
            "type": "object"
          },
          "tools": {
            "additionalProperties": false,
            "title": "Tools",
            "type": "object"
          },
          "workspaces": {
            "additionalProperties": false,
            "title": "Workspaces",
            "type": "object"
          }
        },
        "title": "DryRunProjectionDto",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "additionalProperties": false,
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PreflightReportDto": {
        "additionalProperties": false,
        "properties": {
          "estimated_cost": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Cost"
          },
          "findings": {
            "items": {
              "additionalProperties": false,
              "type": "object"
            },
            "title": "Findings",
            "type": "array"
          },
          "reserved_budget_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reserved Budget Ref"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "unresolved_risks": {
            "items": {
              "type": "string"
            },
            "title": "Unresolved Risks",
            "type": "array"
          }
        },
        "required": [
          "status"
        ],
        "title": "PreflightReportDto",
        "type": "object"
      },
...
      "ProjectSettingsDto": {
        "additionalProperties": false,
        "properties": {
          "budget_policy_id": {
            "title": "Budget Policy Id",
            "type": "string"
          },
          "compute_profile": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Compute Profile"
          },
          "default_model_profile_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Model Profile Id"
          },
          "policy_id": {
            "default": "project-policy",
            "title": "Policy Id",
            "type": "string"
          },
          "project_id": {
            "title": "Project Id",
            "type": "string"
          },
          "team_template_id": {
            "title": "Team Template Id",
            "type": "string"
          },
          "workspace_backend": {
            "title": "Workspace Backend",
            "type": "string"
          }
        },
        "required": [
          "project_id",
          "team_template_id",
          "budget_policy_id",
          "workspace_backend"
        ],
        "title": "ProjectSettingsDto",
        "type": "object"
      },
      "ProtocolSourceDto": {
        "additionalProperties": false,
        "description": "协议文件路径（examples/protocols/ 内，wizard 选择）。",
        "properties": {
          "path": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "path"
        ],
        "title": "ProtocolSourceDto",
        "type": "object"
      },
...
      "ValidationError": {
        "additionalProperties": false,
        "properties": {
          "ctx": {
            "additionalProperties": false,
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Research Console 控制面：配置 / 探测 / 运行 / 审批 / 审计",
    "title": "Research OS Control Plane API",
    "version": "0.4.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/agents/{agent_id}": {
      "patch": {
        "description": "per-Agent 更新（model_binding 等）。\n\nM13 诚实边界：配置目录当前来自 examples/ 契约（只读快照）；运行中\n修改 Model/Tool/semantic 配置必须产生 Manifest Revision / Fork，\n持久化配置修改属 M14（PostgreSQL canonical state）。本端点验证\n变更合法性但不写回（返回 501 说明原因），防止伪持久化。",
        "operationId": "update_agent_agents__agent_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "title": "Agent Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentUpdateDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentSpecDto"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Agent",
        "tags": [
          "team-protocol"
        ]
      }
    },
...
    "/projects/{project_id}/agents": {
      "get": {
        "description": "Agent 实例目录（Role 的配置实例；可独立绑定模型）。",
        "operationId": "list_agents_projects__project_id__agents_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AgentSpecDto"
                  },
                  "title": "Response List Agents Projects  Project Id  Agents Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Agents",
        "tags": [
          "team-protocol"
        ]
      }
    },
    "/projects/{project_id}/compile": {
      "post": {
        "description": "compile + preflight（正式组合入口；预算预留仅发生在 preflight 内）。",
        "operationId": "compile_and_preflight_endpoint_projects__project_id__compile_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProtocolSourceDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreflightReportDto"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Compile And Preflight Endpoint",
        "tags": [
          "team-protocol"
        ]
      }
    },
    "/projects/{project_id}/dry-run": {
      "post": {
        "description": "Dry Run：纯只读投影，零 Research side effect。\n\n不启动 Agent、不调用 Research Tool、不执行 Experiment、不写长期\nMemory、不 reserve budget（`dry_run_projection` 是纯函数）。",
        "operationId": "dry_run_endpoint_projects__project_id__dry_run_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProtocolSourceDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DryRunProjectionDto"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Dry Run Endpoint",
        "tags": [
          "team-protocol"
        ]
      }
    },
    "/projects/{project_id}/preflight": {
      "post": {
        "description": "preflight 复检（编译成功后执行）。",
        "operationId": "preflight_endpoint_projects__project_id__preflight_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProtocolSourceDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreflightReportDto"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preflight Endpoint",
        "tags": [
          "team-protocol"
        ]
      }
    },
...
    "/projects/{project_id}/settings": {
      "get": {
        "description": "项目设置视图（wizard 默认项目；M14 后持久化）。",
        "operationId": "get_project_settings_projects__project_id__settings_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectSettingsDto"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Project Settings",
        "tags": [
          "team-protocol"
        ]
      }
    },
    "/protocols/validate": {
      "post": {
        "description": "协议校验：编译（不做任何副作用）。",
        "operationId": "validate_protocol_protocols_validate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProtocolSourceDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompileResultDto"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Validate Protocol",
        "tags": [
          "team-protocol"
        ]
      }
    },
...
  }
}